requirements.txt
pyserial
argparse
rich
opencv-python
mss
dxcam
simple_pid
pynput
pywin32
pypiwin32
tk
cuda-python==12.6.0
onnx
onnxruntime
onnxsim
PyYAML
matplotlib
nvidia-pyindex

PyQt6
packaging